requirements.txt
tweepy==4.14.0
pandas==2.2.2
numpy==1.26.4
vaderSentiment==3.3.2
scikit-learn==1.4.2
networkx==3.3
matplotlib==3.8.4
python-dotenv==1.0.1
streamlit==1.35.0
textblob==0.18.0.post0
joblib==1.4.2
